burgerstore/requirements.txt
Django==3.2
django-debug-toolbar==3.2.1
requests==2.27.1
djangorestframework==3.13.1
django-debug-toolbar==3.2.1
environs==9.3.2
geopy==2.2.0
Pillow==8.2.0
django-phonenumber-field==6.1.0
phonenumbers==8.12.43
dj-database-url==0.5.0
rollbar==0.16.2
psycopg2-binary==2.9.3
django-admin-sortable2==2.0.1
